pyproject.toml
[project]
name = "benchmarks"
version = "0.1.0"
description = "FriendBench + PressBench: AI benchmark monorepo"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.83.0",
    "click>=8.2.1",
    "fastapi>=0.115.0",
    "google-genai>=1.64.0",
    "inspect-ai>=0.3",
    "openai>=2.22.0",
    "pyyaml>=6.0.3",
    "uvicorn>=0.34.0",
    "xai-sdk>=1.7.0",
]

[project.scripts]
bench = "benchmarks.cli:bench"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["benchmarks"]
